clean: 打刻種類ごとにコンポーネント分け
